Deprecation annotation and tag
A deprecated token such as this function should be annotated with the
java.lang.Deprecated annotation so that compiler warnings may arise and
IDE's such as Android Studio may highlight it as such. The oracle and
google spec dictate that additionally a token should be tagged in the
javadoc with a reference to the replacement, if any.