A charsequence.toString() is never null
Modified this code so it still reflects the old test. However,
appInfo.loadLabel(pm) will (currently) not be able to return a null
value, as the name or packagename is returned if the label is null. That
said, a manufacturer could break this function.

Charsequence#toString on a null value raises a NullPointerException