Improved Branch request handling in slow network
On slow networks If the hostname resolves to multiple IP addresses, URL
connection will attempt to connect the next if one connection fails.
This may result is multiple timeouts elapse before the connect attempt
throws an exception. In order to make sure the Branch callback is
called with in the timeout specified we resolve the host first and then
try only with first resolved ip.
   This will ensure that Branch Callbacks are retuned in the time out
specified (Default is 3sec sec and 3 attempts. so 9 sec max).
 This fix will solve the issue “Branch init callback returned after
long wait on slow networks”