Fixing Activity life cycle re init issue
`getAutoInstance()` or `getAutoTestInstance()` are supposed to be
called only from `Application#onCreate()`. After this Branch instance
can be accessed anywhere using `getInstance()` method.  This fix the
issue of  reinitializing the activity life cycle observer if
`getAutoInstance()` is called  to get instance accidently which will
lead to issues with Branch session management