Fix Maintain DataType In Metadata
ContentMetadata in BUO was  not maintaining the datatype when it is
converted to JSON format. This causes the arrays in content metadata
misplaced in BUO.

Backward compatibility : Both  Android and iOS reads the data in
primitive type or convert the stringified value to original type. So
this is fully backward compatible