Update consumer Proguard configuration
There needs to be a proguard config for the consumer that reflects what projects importing this library will need. In this case, we need to tell these projects it is okay to strip references to Huawei if it cannot compute its usage due to it not being in the classpath.

The motivation behind this change is from recent changes found in Android Gradle Plugin 8. R8 has become more strict and requires projects add this `dontwarn` rule to ensure it is okay to strip out code that is referenced via the `compileOnly` directive but otherwise aren't part of the classpath.